added ability to get session via api
